Add new custom hook
